python/requirements.txt
Authlib==1.2.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
requests==2.31.0
urllib3==2.0.7
validators==0.22.0
weaviate-client==3.25.1
Werkzeug==3.0.1
